Musicosa Runner/requirements.txt
peewee~=3.17.6
validators~=0.34.0
playwright~=1.47.0
yt-dlp~=2024.8.6
typed-ffmpeg~=2.0.0